refactor: orchestrate loader lock from HomeHero; address review #41
Address code review on PR #41:

- PageLoader: cleanup uses removeAttribute("aria-busy") for symmetry
  with inert, leaving a clean DOM instead of a permanent aria-busy="false".
- Move the scroll-lock + inert/aria-busy into a new engagePageLock() helper
  in lib/pageLoader.ts, owned by the HomeHero orchestrator. HomeHero now reads
  and writes the session flag and engages the lock in one effect, so the lock
  no longer depends on child-before-parent effect ordering (Suspense-safe).
  PageLoader becomes purely presentational. The lock releases on the loader's
  AnimatePresence onExitComplete (after the slide-up), preserving the prior
  unmount-timed release; release is idempotent with an unmount safety net.
- Remove byte-identical duplicate images/logo-hearts-white.png (dead weight;
  the served copy in public/images/ is retained).
